States Stuff
In Detail:
1. Added Thessaly, adjusted a few greek states to avoid Bordergore and to their actual shapes, and renamed a few of them
2. Made Constantinople the Capital, and added a demilitarized zone as part of the free trade area (might use it on the focus)
3. Changed Diyarbakir to Amed, the Kurdish name
4. Added a portrait for Greece
5. Revambed the Dynamic naming mechanic, and added the following dynamic names:
Turkish and Greek names for: Mugla, Izmir, and Bursa
Turkish, Armenian and Kurdish names for: Van, Adana, Diyarbakir and Sivas.

Note: You made an update while I had already modified some stuff. There is a small chance of some building errors. If that is the case, use nudge to fix them.